% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcompplot.R
\name{extract_data}
\alias{extract_data}
\title{Extracts relevant data from the model object to plot.}
\usage{
extract_data(model)
}
\arguments{
\item{model}{A mixOmics model object.}
}
\description{
Extracts relevant data from the model object to plot.
}
