% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{extractor}
\alias{extractor}
\title{Extracts relevant data from the model object to plot.}
\usage{
extractor(model, comps = 1:2)
}
\arguments{
\item{model}{A mixOmics model object.}
}
\description{
Extracts relevant data from the model object to plot.
}
\author{
Casey Shannon
}
