% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcompplot.R
\name{ggcompplot}
\alias{ggcompplot}
\alias{ggcompplot.pca}
\alias{ggcompplot.DA}
\alias{ggcompplot.sgccda}
\title{Draws component plots given a mixOmics plot object.}
\usage{
ggcompplot(model, ...)

\method{ggcompplot}{pca}(model, col = NULL, ...)

\method{ggcompplot}{DA}(model, ...)

\method{ggcompplot}{sgccda}(model, ...)
}
\arguments{
\item{model}{A mixOmics model object.}
}
\description{
Draws component plots given a mixOmics plot object.
}
\author{
Casey Shannon
}
