% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvarplot.R
\name{ggvarplot}
\alias{ggvarplot}
\alias{ggvarplot.pca}
\alias{ggvarplot.DA}
\alias{ggvarplot.sgccda}
\title{Draws variate plots given a mixOmics plot object.}
\usage{
ggvarplot(model, topn = 10)

\method{ggvarplot}{pca}(model, topn = 10)

\method{ggvarplot}{DA}(model, topn = 10)

\method{ggvarplot}{sgccda}(model, topn = 10)
}
\arguments{
\item{model}{A mixOmics model object.}

\item{topn}{How many features to visualize the loading vectors of. Ordered by absolute corrlation with the components.}
}
\description{
Draws variate plots given a mixOmics plot object.
}
\author{
Casey Shannon
}
