% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbiplot.R
\name{ggbiplot}
\alias{ggbiplot}
\alias{ggbiplot.pca}
\alias{ggbiplot.DA}
\alias{ggbiplot.sgccda}
\title{Draws a biplot given a mixOmics plot object.}
\usage{
ggbiplot(model, ...)

\method{ggbiplot}{pca}(model, col = NULL, ...)

\method{ggbiplot}{DA}(model, ...)

\method{ggbiplot}{sgccda}(model, ...)
}
\arguments{
\item{model}{A mixOmics model object.}
}
\description{
Draws a biplot given a mixOmics plot object.
}
\author{
Casey Shannon
}
