% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbiplot.R
\name{ggbiplot}
\alias{ggbiplot}
\title{Draws a biplot given a mixOmics plot object.}
\usage{
ggbiplot(model, col = NULL, topn = 10, ...)
}
\arguments{
\item{model}{A mixOmics model object.}
}
\description{
Draws a biplot given a mixOmics plot object.
}
\author{
Casey Shannon
}
