% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbiplot.R
\name{ggbiplot}
\alias{ggbiplot}
\alias{ggbiplot.pca}
\alias{ggbiplot.DA}
\alias{ggbiplot.sgccda}
\title{Draws a biplot given a mixOmics plot object.}
\usage{
ggbiplot(model, ...)

\method{ggbiplot}{pca}(model, comps = 1:2, col = NULL, ...)

\method{ggbiplot}{DA}(model, comps = 1:2, ...)

\method{ggbiplot}{sgccda}(model, comps = 1:2, ...)
}
\arguments{
\item{model}{A mixOmics model object.}

\item{...}{Optional. Parameters passed to ggvarplot.}

\item{comps}{The components to visualize.}

\item{col}{Optional. A vector. Labels to colour points by.}
}
\description{
Draws a biplot given a mixOmics plot object.
}
\author{
Casey Shannon
}
