% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcompplot.R
\name{ggcompplot}
\alias{ggcompplot}
\title{Draws component and variate plots given a mixOmics plot object, optionally combining them.}
\usage{
ggcompplot(model, topn = NULL, combined = F)
}
\arguments{
\item{model}{A mixOmics model object.}

\item{topn}{How many features to visualize the loading vectors of. Ordered by absolute corrlation with the components.}
}
\description{
Draws component and variate plots given a mixOmics plot object, optionally combining them.
}
\examples{

model <- mixOmics::splsda(X, Y, keepX = c(10, 10))
ggcompplot(model)
}
