% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcompplot.R
\name{ggcompplot}
\alias{ggcompplot}
\alias{ggcompplot.pca}
\alias{ggcompplot.DA}
\alias{ggcompplot.sgccda}
\title{Draws component plots given a mixOmics plot object.}
\usage{
ggcompplot(model, ...)

\method{ggcompplot}{pca}(model, comps = 1:2, col = NULL, ...)

\method{ggcompplot}{DA}(model, comps = 1:2, ...)

\method{ggcompplot}{sgccda}(model, comps = 1:2, ...)
}
\arguments{
\item{model}{A mixOmics model object.}

\item{comps}{The components to visualize.}

\item{col}{Optional. A vector. Labels to colour points by.}
}
\description{
Draws component plots given a mixOmics plot object.
}
\author{
Casey Shannon
}
